+++ lib/src/l10n/l10n_helper.g.dart
@@ -1,0 +1,24 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'l10n_helper.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$appLocaleHash() => r'e3cc272e4d0f400d06f65e631d5ce4b36778b073';
+
+/// See also [AppLocale].
+@ProviderFor(AppLocale)
+final appLocaleProvider = NotifierProvider<AppLocale, Locale?>.internal(
+  AppLocale.new,
+  name: r'appLocaleProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$appLocaleHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$AppLocale = Notifier<Locale?>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member