+++ lib/src/routes/routes.g.dart
@@ -43,6 +43,13 @@ RouteBase get $rootRoute => GoRouteData.$route(
           path: 'settings',
           name: 'settings',
           factory: $SettingsRouteExtension._fromState,
+          routes: [
+            GoRouteData.$route(
+              path: 'LICENSE',
+              name: 'license',
+              factory: $ProjectLicenseRouteExtension._fromState,
+            ),
+          ],
         ),
       ],
     );
@@ -154,6 +161,24 @@ extension $SettingsRouteExtension on SettingsRoute {
 
   String get location => GoRouteData.$location(
         '/settings',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+extension $ProjectLicenseRouteExtension on ProjectLicenseRoute {
+  static ProjectLicenseRoute _fromState(GoRouterState state) =>
+      ProjectLicenseRoute();
+
+  String get location => GoRouteData.$location(
+        '/settings/LICENSE',
       );
 
   void go(BuildContext context) => context.go(location);