+++ lib/src/ui/app.dart
@@ -1,4 +1,3 @@
-import 'package:dreambook/src/l10n/app_localizations.dart';
 import 'package:dreambook/src/l10n/l10n_helper.dart';
 import 'package:dreambook/src/routes/routes.dart';
 import 'package:dreambook/src/ui/theme.dart';
@@ -18,6 +17,7 @@ class App extends ConsumerWidget {
       routerConfig: ref.watch(appRoutesProvider),
       supportedLocales: AppLocalizations.supportedLocales,
       localizationsDelegates: AppLocalizations.localizationsDelegates,
+      locale: ref.watch(appLocaleProvider),
     );
   }
 }